requirements.txt
pdf2image
pyperclip
python-dotenv
python-gyazo
tqdm
